fix: address PR review findings for uptime/lastUser feature
- Add lastUser and uptimeSeconds to device list API select and response
  mapping so overview cards populate when navigating from the list
- Log warning when host.BootTime() fails instead of silently swallowing
- Add .max(255) to lastUser and .min(0) to uptime Zod validation
- Extract duplicated formatUptime into shared lib/utils.ts
