+++ apps/web/src/components/devices/DeviceInfoTab.tsx
@@ -1,6 +1,7 @@
 import { useCallback, useEffect, useState } from 'react';
 import { Monitor, Cpu, HardDrive, MemoryStick, Shield, Tag, Info, ListChecks, Pencil, Check, X } from 'lucide-react';
 import { fetchWithAuth } from '../../stores/auth';
+import { formatUptime } from '../../lib/utils';
 
 type DeviceInfoTabProps = {
   deviceId: string;
@@ -69,16 +70,6 @@ function formatDate(dateString: string | null | undefined): string {
     hour: '2-digit',
     minute: '2-digit',
   });
-}
-
-function formatUptime(seconds: number | null | undefined): string {
-  if (seconds === null || seconds === undefined || seconds < 0) return '—';
-  const days = Math.floor(seconds / 86400);
-  const hours = Math.floor((seconds % 86400) / 3600);
-  const minutes = Math.floor((seconds % 3600) / 60);
-  if (days > 0) return `${days}d ${hours}h ${minutes}m`;
-  if (hours > 0) return `${hours}h ${minutes}m`;
-  return `${minutes}m`;
 }
 
 function InfoRow({ label, value }: { label: string; value: string }) {
